+++ src/pint/models/stand_alone_psr_binaries/binary_generic.py
@@ -1,9 +1,10 @@
 """Parent class for internal binary models."""
-import logging
 
 import astropy.constants as c
 import astropy.units as u
 import numpy as np
+
+from loguru import logger as log
 
 try:
     from erfa import DAYSEC as SECS_PER_DAY
@@ -12,8 +13,6 @@ except ImportError:
 
 from pint import Tsun, ls
 from pint.models.stand_alone_psr_binaries.binary_orbits import OrbitPB
-
-log = logging.getLogger(__name__)
 
 SECS_PER_JUL_YEAR = SECS_PER_DAY * 365.25