+++ src/pint/models/timing_model.py
@@ -29,7 +29,6 @@ See :ref:`Timing Models` for more details on how PINT's timing models work.
 import abc
 import copy
 import inspect
-import logging
 from collections import OrderedDict, defaultdict
 from functools import wraps
 from warnings import warn
@@ -40,6 +39,7 @@ import numpy as np
 from astropy import log
 from astropy.utils.decorators import lazyproperty
 from scipy.optimize import brentq
+from loguru import logger as log
 
 import pint
 from pint.models.parameter import (
@@ -58,7 +58,6 @@ from pint.phase import Phase
 from pint.toa import TOAs
 from pint.utils import PrefixError, interesting_lines, lines_of, split_prefixed_name
 
-log = logging.getLogger(__name__)
 
 __all__ = [
     "DEFAULT_ORDER",
@@ -1375,12 +1374,15 @@ class TimingModel:
         """
         from . import jump
 
+        # Because JUMPS are handled serially, we need to do everything
+        # via index order and *not* by group_by("obs")
+        ix_tab = toas.table[np.argsort(toas.table["index"])]
         # check if any TOAs are jumped
-        jumped = ["jump" in flag_dict.keys() for flag_dict in toas.table["flags"]]
+        jumped = ["jump" in flag_dict.keys() for flag_dict in ix_tab["flags"]]
         if not any(jumped):
             log.info("No jump flags to process from .tim file")
             return None
-        for flag_dict in toas.table["flags"][jumped]:
+        for flag_dict in ix_tab["flags"][jumped]:
             # add PhaseJump object if model does not have one already
             if "PhaseJump" not in self.components:
                 log.info("PhaseJump component added")
@@ -1389,22 +1391,20 @@ class TimingModel:
                 self.add_component(a)
                 self.remove_param("JUMP1")
             # take jumps in TOA table and add them as parameters to the model
-            for num in flag_dict["jump"]:
-                if "JUMP" + str(num) not in self.params:
-                    param = maskParameter(
-                        name="JUMP",
-                        index=num,
-                        key="-tim_jump",
-                        key_value=num,
-                        value=0.0,
-                        units="second",
-                        uncertainty=0.0,
-                    )
-                    self.add_param_from_top(param, "PhaseJump")
-                    getattr(self, param.name).frozen = False
-                flag_dict["tim_jump"] = str(
-                    num
-                )  # this is the value select_toa_mask uses
+            num = flag_dict["jump"]
+            if "JUMP" + str(num) not in self.params:
+                param = maskParameter(
+                    name="JUMP",
+                    index=num,
+                    key="-tim_jump",
+                    key_value=num,
+                    value=0.0,
+                    units="second",
+                    uncertainty=0.0,
+                )
+                self.add_param_from_top(param, "PhaseJump")
+                getattr(self, param.name).frozen = False
+            flag_dict["tim_jump"] = str(num)  # this is the value select_toa_mask uses
         self.components["PhaseJump"].setup()
 
     def delete_jump_and_flags(self, toa_table, jump_num):