+++ src/pint/models/troposphere_delay.py
@@ -1,12 +1,10 @@
 """Delay due to Earth's troposphere"""
-import logging
-from warnings import warn
-
 import astropy.constants as const
 import astropy.units as u
 import numpy as np
 import scipy.interpolate
 from astropy.coordinates import AltAz, SkyCoord
+from loguru import logger as log
 
 import pint.utils as ut
 from pint.models.parameter import boolParameter
@@ -14,8 +12,6 @@ from pint.models.timing_model import DelayComponent
 from pint.observatory import get_observatory
 from pint.observatory.topo_obs import TopoObs
 from pint.toa_select import TOASelect
-
-log = logging.getLogger(__name__)
 
 
 class TroposphereDelay(DelayComponent):