+++ src/pint/observatory/topo_obs.py
@@ -1,11 +1,11 @@
 """Ground-based fixed observatories."""
-import logging
 import os
 
 import astropy.constants as c
 import astropy.units as u
 import numpy
 from astropy.coordinates import EarthLocation
+from loguru import logger as log
 
 from pint import JD_MJD
 from pint.config import runtimefile
@@ -15,8 +15,6 @@ from pint.observatory.clock_file import ClockFile
 from pint.pulsar_mjd import Time
 from pint.solar_system_ephemerides import get_tdb_tt_ephem_geocenter, objPosVel_wrt_SSB
 from pint.utils import has_astropy_unit
-
-log = logging.getLogger(__name__)
 
 
 class TopoObs(Observatory):
@@ -253,12 +251,12 @@ class TopoObs(Observatory):
             )
             tt2tai = 32.184 * 1e6 * u.us
             if self._bipm_clock is None:
-                try:
-                    log.info(
-                        "Observatory {0}, loading BIPM clock file \n\t{1}".format(
-                            self.name, self.bipm_fullpath
-                        )
+                log.info(
+                    "Observatory {0}, loading BIPM clock file \n\t{1}".format(
+                        self.name, self.bipm_fullpath
                     )
+                )
+                try:
                     self._bipm_clock = ClockFile.read(
                         self.bipm_fullpath, format="tempo2"
                     )