+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project, at least loosely, adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added computation of other Kopeikin solutions (`solutions = model.alternative_solutions()`)
+- Added computation of extra parameters in gridding
+- Added gridding based on tuples of parameters (not just regular mesh)
+- Added passing of extra parameters to gridding fitter
+### Fixed
+### Changed
+- Changed logging to use `loguru`
+
+
 ## [0.8.5] 2022-02-24
 ### Added
 - Added support for Chandra and Swift event data in photonphase.py